+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.4.1 (TBD)
 
 - Fixed issue where switching to The Judge on a new floor from a weapon with the same base ammo type would break The Judge's ammo UI rendering
+- Added "Preallocate Heap Memory" option
 
 ## 1.4.0 (2025-06-06)